requirements.txt
streamlit>=1.32
phidata
python-dotenv
duckduckgo-search
openai
